Created registration an login functions

+++ base/views.py
@@ -1,18 +1,82 @@
 from django.http import HttpResponse
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
-from . models import Room
+from django.db.models import  Q
+from . models import *
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate,login,logout
 from .forms import RoomForm
+from django.contrib import messages
 
 # Create your views here.
 
+
+def login_page(request):
+    page='login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method =='POST':
+        username=request.POST.get('username').lower()
+        password=request.POST.get('password')
+
+        try:
+            user=User.objects.get(username=username)
+        except User.DoesNotExist:
+            messages.error(request,'User does Not Exist')
+
+        user=authenticate(request, username=username,password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request,'User not found')
+
+    context={'page':page}
+    return render(request,'base/login_register.html', context)
+
+def register_user(request):
+    page='register'
+    form=UserCreationForm()
+    if request.method =='POST':
+        form=UserCreationForm(request.POST)
+        if form.is_valid():
+            user=form.save(commit=False)
+            user.username=user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+
+    context={'page':page,'form':form}
+    return render(request,'base/login_register.html', context)
+
+
 def home(request):
-    rooms= Room.objects.all()
+
+    q=request.GET.get('q') if request.GET.get('q') != None else ''
+    rooms= Room.objects.filter(
+        Q(topic__name__icontains=q)|
+        Q(name__icontains=q)|
+        Q(description__icontains=q)
+        )
+    topics=Topic.objects.all()
+    room_count=rooms.count()
 
     context={
-        'rooms':rooms
+        'rooms':rooms,
+        'topics':topics,
+        'room_count':room_count,
+        
     }
     return render(request,'base/home.html',context)
+
+
+def logout_user(request):
+    logout(request)
+    return redirect('home')
 
 
 def room(request,pk):
@@ -22,6 +86,7 @@ def room(request,pk):
     }
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method =='POST':
@@ -33,9 +98,13 @@ def createRoom(request):
     context={'form':form}
     return render(request,'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+
+    if request.user != room.host :
+        return HttpResponse('You are not allowed to update this')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -48,8 +117,13 @@ def updateRoom(request, pk):
     context={'form':form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room= Room.objects.get(id=pk)
+
+    if request.user != room.host :
+        return HttpResponse('You are not allowed to delete this')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
